+++ src/braket/jobs/hybrid_job.py
@@ -46,12 +46,12 @@ from braket.jobs.quantum_job_creation import _generate_default_job_name
 def hybrid_job(
     *,
     device: str | None,
-    include_modules: str | ModuleType | Iterable[str | ModuleType] = None,
-    dependencies: str | Path | None = None,
+    include_modules: str | ModuleType | Iterable[str | ModuleType] | None = None,
+    dependencies: str | Path | list[str] | None = None,
     local: bool = False,
     job_name: str | None = None,
     image_uri: str | None = None,
-    input_data: str | dict | S3DataSourceConfig = None,
+    input_data: str | dict | S3DataSourceConfig | None = None,
     wait_until_complete: bool = False,
     instance_config: InstanceConfig | None = None,
     distribution: str | None = None,
@@ -61,15 +61,15 @@ def hybrid_job(
     stopping_condition: StoppingCondition | None = None,
     output_data_config: OutputDataConfig | None = None,
     aws_session: AwsSession | None = None,
-    tags: dict[str, str] = None,
+    tags: dict[str, str] | None = None,
     logger: Logger = getLogger(__name__),
 ) -> Callable:
     """Defines a hybrid job by decorating the entry point function. The job will be created
     when the decorated function is called.
 
     The job created will be a `LocalQuantumJob` when `local` is set to `True`, otherwise an
-    `AwsQuantumJob. The following parameters will be ignored when running a job with
-    `local` set to True: `wait_until_complete`, `instance_config`, `distribution`,
+    `AwsQuantumJob`. The following parameters will be ignored when running a job with
+    `local` set to `True`: `wait_until_complete`, `instance_config`, `distribution`,
     `copy_checkpoints_from_job`, `stopping_condition`, `tags`, and `logger`.
 
     Args:
@@ -80,15 +80,17 @@ def hybrid_job(
             When using embedded simulators, you may provide the device argument as string of the
             form: "local:<provider>/<simulator_name>" or `None`.
 
-        include_modules (str | ModuleType | Iterable[str | ModuleType]): Either a
+        include_modules (str | ModuleType | Iterable[str | ModuleType] | None): Either a
             single module or module name or a list of module or module names referring to local
             modules to be included. Any references to members of these modules in the hybrid job
-            algorithm code will be serialized as part of the algorithm code. Default value `[]`
+            algorithm code will be serialized as part of the algorithm code. Default: `[]`
 
-        dependencies (str | Path | None): Path (absolute or relative) to a requirements.txt
-            file to be used for the hybrid job.
+        dependencies (str | Path | list[str] | None): Path (absolute or relative) to a
+            requirements.txt file, or alternatively a list of strings, with each string being a
+            `requirement specifier <https://pip.pypa.io/en/stable/reference/requirement-specifiers/
+            #requirement-specifiers>`_, to be used for the hybrid job.
 
-        local (bool): Whether to use local mode for the hybrid job. Default `False`
+        local (bool): Whether to use local mode for the hybrid job. Default: `False`
 
         job_name (str | None): A string that specifies the name with which the job is created.
             Allowed pattern for job name: `^[a-zA-Z0-9](-*[a-zA-Z0-9]){0,50}$`. Defaults to
@@ -96,12 +98,12 @@ def hybrid_job(
 
         image_uri (str | None): A str that specifies the ECR image to use for executing the job.
             `retrieve_image()` function may be used for retrieving the ECR image URIs
-            for the containers supported by Braket. Default = `<Braket base image_uri>`.
+            for the containers supported by Braket. Default: `<Braket base image_uri>`.
 
-        input_data (str | Dict | S3DataSourceConfig): Information about the training
+        input_data (str | dict | S3DataSourceConfig | None): Information about the training
             data. Dictionary maps channel names to local paths or S3 URIs. Contents found
             at any local paths will be uploaded to S3 at
-            f's3://{default_bucket_name}/jobs/{job_name}/data/{channel_name}. If a local
+            f's3://{default_bucket_name}/jobs/{job_name}/data/{channel_name}'. If a local
             path, S3 URI, or S3DataSourceConfig is provided, it will be given a default
             channel name "input".
             Default: {}.
@@ -111,23 +113,23 @@ def hybrid_job(
             local mode. Default: `False`.
 
         instance_config (InstanceConfig | None): Configuration of the instance(s) for running the
-            classical code for the hybrid job. Defaults to
+            classical code for the hybrid job. Default:
             `InstanceConfig(instanceType='ml.m5.large', instanceCount=1, volumeSizeInGB=30)`.
 
         distribution (str | None): A str that specifies how the job should be distributed.
             If set to "data_parallel", the hyperparameters for the job will be set to use data
-            parallelism features for PyTorch or TensorFlow. Default: None.
+            parallelism features for PyTorch or TensorFlow. Default: `None`.
 
         copy_checkpoints_from_job (str | None): A str that specifies the job ARN whose
             checkpoint you want to use in the current job. Specifying this value will copy
             over the checkpoint data from `use_checkpoints_from_job`'s checkpoint_config
             s3Uri to the current job's checkpoint_config s3Uri, making it available at
-            checkpoint_config.localPath during the job execution. Default: None
+            checkpoint_config.localPath during the job execution. Default: `None`
 
         checkpoint_config (CheckpointConfig | None): Configuration that specifies the
             location where checkpoint data is stored.
-            Default: CheckpointConfig(localPath='/opt/jobs/checkpoints',
-            s3Uri=f's3://{default_bucket_name}/jobs/{job_name}/checkpoints').
+            Default: `CheckpointConfig(localPath='/opt/jobs/checkpoints',
+            s3Uri=f's3://{default_bucket_name}/jobs/{job_name}/checkpoints')`.
 
         role_arn (str | None): A str providing the IAM role ARN used to execute the
             script. Default: IAM role returned by AwsSession's `get_default_jobs_role()`.
@@ -138,8 +140,8 @@ def hybrid_job(
 
         output_data_config (OutputDataConfig | None): Specifies the location for the output of
             the job.
-            Default: OutputDataConfig(s3Path=f's3://{default_bucket_name}/jobs/{job_name}/data',
-            kmsKeyId=None).
+            Default: `OutputDataConfig(s3Path=f's3://{default_bucket_name}/jobs/{job_name}/data',
+            kmsKeyId=None)`.
 
         aws_session (AwsSession | None): AwsSession for connecting to AWS Services.
             Default: AwsSession()
@@ -148,14 +150,21 @@ def hybrid_job(
             Default: {}.
 
         logger (Logger): Logger object with which to write logs, such as task statuses
-            while waiting for task to be in a terminal state. Default is `getLogger(__name__)`
-    """
-    aws_session = aws_session or AwsSession()
-    _validate_python_version(aws_session, image_uri)
+            while waiting for task to be in a terminal state. Default: `getLogger(__name__)`
 
-    def _hybrid_job(entry_point):
+    Returns:
+        Callable: the callable for creating a Hybrid Job.
+    """
+    _validate_python_version(image_uri, aws_session)
+
+    def _hybrid_job(entry_point: Callable) -> Callable:
         @functools.wraps(entry_point)
-        def job_wrapper(*args, **kwargs):
+        def job_wrapper(*args, **kwargs) -> Callable:
+            """
+            The job wrapper.
+            Returns:
+                Callable: the callable for creating a Hybrid Job.
+            """
             with _IncludeModules(include_modules), tempfile.TemporaryDirectory(
                 dir="", prefix="decorator_job_"
             ) as temp_dir:
@@ -171,7 +180,7 @@ def hybrid_job(
                     entry_point_file.write(template)
 
                 if dependencies:
-                    shutil.copy(Path(dependencies).resolve(), temp_dir_path / "requirements.txt")
+                    _process_dependencies(dependencies, temp_dir_path)
 
                 job_args = {
                     "device": device or "local:none/none",
@@ -209,8 +218,9 @@ def hybrid_job(
     return _hybrid_job
 
 
-def _validate_python_version(aws_session: AwsSession, image_uri: str | None):
+def _validate_python_version(image_uri: str | None, aws_session: AwsSession | None = None) -> None:
     """Validate python version at job definition time"""
+    aws_session = aws_session or AwsSession()
     # user provides a custom image_uri
     if image_uri and image_uri not in built_in_images(aws_session.region):
         print(
@@ -231,6 +241,16 @@ def _validate_python_version(aws_session: AwsSession, image_uri: str | None):
                 f"Client is running Python {sys.version_info.major}.{sys.version_info.minor} "
                 f"locally, but container uses Python {major_version}.{minor_version}."
             )
+
+
+def _process_dependencies(dependencies: str | Path | list[str], temp_dir: Path) -> None:
+    if isinstance(dependencies, (str, Path)):
+        # requirements file
+        shutil.copy(Path(dependencies).resolve(), temp_dir / "requirements.txt")
+    else:
+        # list of packages
+        with open(temp_dir / "requirements.txt", "w") as f:
+            f.write("\n".join(dependencies))
 
 
 class _IncludeModules:
@@ -256,10 +276,7 @@ class _IncludeModules:
 
 def _serialize_entry_point(entry_point: Callable, args: tuple, kwargs: dict) -> str:
     """Create an entry point from a function"""
-
-    def wrapped_entry_point():
-        """Partial function wrapping entry point with given parameters"""
-        return entry_point(*args, **kwargs)
+    wrapped_entry_point = functools.partial(entry_point, *args, **kwargs)
 
     try:
         serialized = cloudpickle.dumps(wrapped_entry_point)
@@ -277,7 +294,7 @@ def _serialize_entry_point(entry_point: Callable, args: tuple, kwargs: dict) -> 
     )
 
 
-def _log_hyperparameters(entry_point: Callable, args: tuple, kwargs: dict):
+def _log_hyperparameters(entry_point: Callable, args: tuple, kwargs: dict) -> dict:
     """Capture function arguments as hyperparameters"""
     signature = inspect.signature(entry_point)
     bound_args = signature.bind(*args, **kwargs)
@@ -318,11 +335,11 @@ def _sanitize(hyperparameter: Any) -> str:
     # max allowed length for a hyperparameter is 2500
     if len(sanitized) > 2500:
         # show as much as possible, including the final 20 characters
-        return f"{sanitized[:2500-23]}...{sanitized[-20:]}"
+        return f"{sanitized[:2500 - 23]}...{sanitized[-20:]}"
     return sanitized
 
 
-def _process_input_data(input_data):
+def _process_input_data(input_data: dict) -> list[str]:
     """
     Create symlinks to data
 
@@ -336,12 +353,12 @@ def _process_input_data(input_data):
     if not isinstance(input_data, dict):
         input_data = {"input": input_data}
 
-    def matches(prefix):
+    def matches(prefix: str) -> list[str]:
         return [
             str(path) for path in Path(prefix).parent.iterdir() if str(path).startswith(str(prefix))
         ]
 
-    def is_prefix(path):
+    def is_prefix(path: str) -> bool:
         return len(matches(path)) > 1 or not Path(path).exists()
 
     prefix_channels = set()
@@ -349,7 +366,7 @@ def _process_input_data(input_data):
     file_channels = set()
 
     for channel, data in input_data.items():
-        if AwsSession.is_s3_uri(str(data)):
+        if AwsSession.is_s3_uri(str(data)) or isinstance(data, S3DataSourceConfig):
             channel_arg = f'channel="{channel}"' if channel != "input" else ""
             print(
                 "Input data channels mapped to an S3 source will not be available in "